Corrected hours of stores, added "Open" or "Closed" functionality,  re-added user current location icon, and commented code
Corrected hours of stores by adding list inside with the hours of the store by the days of the week. Utilized getDayOfWeek funciton to get the current day of the week.

Depending on current user's time and the stores time, will display if the store is open or closed with isOpenNow function

Re-added default icon to show user's current location

Commented code

, added "Open" or "Closed" functionality,  re-added user current location icon, and commented code